Update README.md
First: Thanks a lot for this nice project!! 

It is an good add-on to my customized GitHub-Startpage: https://github.com/OllisGit.

At had some issues to setup the Fork, e.g. I entered the ID like it is in the plugins.octoprint.org (needs to be lowercase). And I was a little bit confused about the order of the action-script.
So, that was the reason why I added an addition section to the readme.

Have a nice day,
Olli